# openapi.yaml
openapi: 3.0.4
info:
  title: Message API
  description: Sample Description
  version: 1.0.12
servers:
  - url: http://localhost:8080
tags:
  - name: messages
paths:
  /{channelID}/messages:
    get:
      summary: Get messages
      description: Get messages
      operationId: getMessages
      parameters: 
        - in: path
          name: channelID
          schema:
            type: integer
          required: true
          description: Channel to search for messages in
        - in: query
          name: Username
          schema:
            type: string
          required: false
          description: Username to **exclude**
        - in: query
          name: Channel
          schema:
            type: string
          required: false
          description: Channel to search in
        - in: query
          name: Since
          schema:
            type: integer
          required: false
          description: Earliest timestamp to search for
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Messages'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Insert a new message
      description: Insert a new message
      operationId: insertMessage
      parameters: 
        - in: path
          name: channelID
          schema:
            type: integer
          required: true
          description: Channel to search for messages in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        '200':
          description: Successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 'b4a72348-19ab-4b7e-bfd8-ac9566cb07da'
        timestamp:
          type: integer
          format: int64
          example: 1746298639518
        username:
          type: string
          example: "testUser"
        channel:
          type: string
          example: "channel1"
        content:
          type: string
          example: "Your message content"
    Messages:
      type: array
      items:
        $ref: '#/components/schemas/Message'
    Error:
      type: string
      example: "An error ocurred"